requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
jwt==1.3.1
Markdown==3.3.7
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.0
